Crate decaf377_ka

Structs

Public

A public key sent to the counterparty in the key agreement protocol.

Secret

A secret key used to perform key agreement using the counterparty’s public key.

SharedSecret

The shared secret derived at the end of the key agreement protocol.

Enums

Error

An error during key agreement.